feat: refactor modules file
