+++ package.json
@@ -11,7 +11,8 @@
     "fs:delete": "node src/fs/delete.js",
     "fs:read": "node src/fs/read.js",
     "cli:env": "RSS_name1=\"value1\" NoRSSenvVar=\"port\" RsS_incorrectPattern=\"asd\" RSS_name2=\"value2\" node src/cli/env.js",
-    "cli:args": "node src/cli/args.js --propName value --someAnotherProp 12312aqe --prop2Name value2"
+    "cli:args": "node src/cli/args.js --propName value --someAnotherProp 12312aqe --prop2Name value2",
+    "modules:cjsToEsm": "node src/modules/cjsToEsm.mjs"
   },
   "repository": {
     "type": "git",
